added methods to season class, and created Reusables module
