+++ lib/games.rb
@@ -1,6 +1,8 @@
 require './lib/details_loader'
 
+
 class Games < DetailsLoader
+  
 
   def initialize(games, teams, game_teams)
     super(games, teams, game_teams)
