+++ lib/stat_tracker.rb
@@ -2,6 +2,7 @@ require 'csv'
 require './lib/csv_loader.rb'
 require './lib/games'
 require './lib/league'
+require './lib/season_stats'
 
 
 class StatTracker
@@ -10,6 +11,7 @@ class StatTracker
   def initialize (games, teams, game_teams)
     @gamezz = Games.new(games, teams, game_teams)
     @league = League.new(games, teams, game_teams)
+    @season_stats = SeasonStats.new(games, teams, game_teams)
     # @csv = CsvLoader.new(games, teams, game_teams)
     # @details = DetailsLoader.new.load_all
     # @games = games
@@ -234,52 +236,53 @@ class StatTracker
 
   # Season Statistics
 
-  def coach_by_team_id #Provides hash of coach names by team id - Helper method for issue #27
-    # Example hash: {3=>"John Tortorella", 6=>"Claude Julien",  5=>"Dan Bylsma",
-    coaches = Hash.new {|h, k| h[k] = {}}
-    games_by_season.each do |season, games|
-      @game_teams.each do |row|
-        if games.include?(row[:game_id])
-          if !coaches[row[:team_id]].keys.include?(season)
-            coaches[row[:team_id]][season] = [row[:head_coach]]
-          elsif !coaches[row[:team_id]][season].include?(row[:head_coach])
-            coaches[row[:team_id]][season] << row[:head_coach]
-          end
-        end
-      end
-    end
-    coaches
-  end
+  # def coach_by_team_id #Provides hash of coach names by team id - Helper method for issue #27
+  #   # Example hash: {3=>"John Tortorella", 6=>"Claude Julien",  5=>"Dan Bylsma",
+  #   coaches = Hash.new {|h, k| h[k] = {}}
+  #   games_by_season.each do |season, games|
+  #     @game_teams.each do |row|
+  #       if games.include?(row[:game_id])
+  #         if !coaches[row[:team_id]].keys.include?(season)
+  #           coaches[row[:team_id]][season] = [row[:head_coach]]
+  #         elsif !coaches[row[:team_id]][season].include?(row[:head_coach])
+  #           coaches[row[:team_id]][season] << row[:head_coach]
+  #         end
+  #       end
+  #     end
+  #   end
+  #   coaches
+  # end
 
-  def team_win_percent_by_season # Provides hash of seasons with array of hashes for team id
-                                  #and win percentage for season - Helper for Issue #27
-    # Example hash: {20132014=> [{:team_id=>1, :win_perc=>50.0}, {:team_id=>4, :win_perc=>40.0}, {:team_id=>26, :win_perc=>100.0}
-    team_win_percent = Hash.new {0}
-    team_by_id.each do |id , team|
-        season_win_percentage(id).each do |season, win|
-          if number_team_games_per_season(id)[season] != (0.0 || 0)
-            if team_win_percent[season] == 0
-              team_win_percent[season] = [{team_id: id, win_perc: win}]
-            else
-              team_win_percent[season] << {team_id: id, win_perc: win}
-          end
-        end
-      end
-    end
-    team_win_percent
-  end
+  # def team_win_percent_by_season # Provides hash of seasons with array of hashes for team id
+  #                                 #and win percentage for season - Helper for Issue #27
+  #   # Example hash: {20132014=> [{:team_id=>1, :win_perc=>50.0}, {:team_id=>4, :win_perc=>40.0}, {:team_id=>26, :win_perc=>100.0}
+  #   team_win_percent = Hash.new {0}
+  #   team_by_id.each do |id , team|
+  #       season_win_percentage(id).each do |season, win|
+  #         if number_team_games_per_season(id)[season] != (0.0 || 0)
+  #           if team_win_percent[season] == 0
+  #             team_win_percent[season] = [{team_id: id, win_perc: win}]
+  #           else
+  #             team_win_percent[season] << {team_id: id, win_perc: win}
+  #         end
+  #       end
+  #     end
+  #   end
+  #   team_win_percent
+  # end
 
 
   def winningest_coach(season) #issue # 17 - FAIL wrong name returns
    # Name of the Coach with the best win percentage for the season
-   highest_percent_wins = team_win_percent_by_season[season.to_i].max_by {|stat| stat[:win_perc]}
-   coach_by_team_id[highest_percent_wins[:team_id]][season.to_i].sample
-
+  #  highest_percent_wins = team_win_percent_by_season[season.to_i].max_by {|stat| stat[:win_perc]}
+  #  coach_by_team_id[highest_percent_wins[:team_id]][season.to_i].sample
+    @season_stats.winningest_coach(season)
   end
 
   def worst_coach(season)#issue # 27 - FAIL 2/3 tests PASS - on fail it is providing diff team/name
-    lowest_percent_wins = team_win_percent_by_season[season.to_i].min_by {|stat| stat[:win_perc]}
-    coach_by_team_id[lowest_percent_wins[:team_id]][season.to_i].sample
+    # lowest_percent_wins = team_win_percent_by_season[season.to_i].min_by {|stat| stat[:win_perc]}
+    # coach_by_team_id[lowest_percent_wins[:team_id]][season.to_i].sample
+    @season_stats.worst_coach(season)
   end
 
   def game_teams_for_game_id(game_id) #Helper method for issue #28, may be able to be used for other season stats
@@ -405,100 +408,100 @@ class StatTracker
       info
   end
 
-  def games_by_season # All game_ids sorted by season - helper method for issue #18
-    # {20122013=> [2012030221, 2012030222, 2012030223, 2012030224,
-    games_by_season = {}
-    @games.values_at(:game_id, :season).each do |game|
-      if games_by_season.include?(game[1])
-        games_by_season[game[1]] << game[0]
-      else
-        games_by_season[game[1]] = [game[0]]
-      end
-    end
-    games_by_season
-  end
+  # def games_by_season # All game_ids sorted by season - helper method for issue #18
+  #   # {20122013=> [2012030221, 2012030222, 2012030223, 2012030224,
+  #   games_by_season = {}
+  #   @games.values_at(:game_id, :season).each do |game|
+  #     if games_by_season.include?(game[1])
+  #       games_by_season[game[1]] << game[0]
+  #     else
+  #       games_by_season[game[1]] = [game[0]]
+  #     end
+  #   end
+  #   games_by_season
+  # end
 
-  def wins_by_team(team_id) # List of every game that was a win for a team - helper method for issue #18
-    # [[2013020252, 16], [2014030166, 16], [2016030151, 16], [2016030152, 16]]
-    # wins = []
-    # @games.each do |row|
-    #   if (row[:away_goals] > row[:home_goals] && row[:away_team_id] == team_id) ||
-    #     (row[:home_goals] < row[:away_goals] && row[:home_team_id] == team_id)
-    #     wins << [row[:game_id], team_id]
-    #   end
-    # end
-    # wins
+  # def wins_by_team(team_id) # List of every game that was a win for a team - helper method for issue #18
+  #   # [[2013020252, 16], [2014030166, 16], [2016030151, 16], [2016030152, 16]]
+  #   # wins = []
+  #   # @games.each do |row|
+  #   #   if (row[:away_goals] > row[:home_goals] && row[:away_team_id] == team_id) ||
+  #   #     (row[:home_goals] < row[:away_goals] && row[:home_team_id] == team_id)
+  #   #     wins << [row[:game_id], team_id]
+  #   #   end
+  #   # end
+  #   # wins
 
-    # Option with full data set, but does not work with current dummy data
-    # (this could be made dynamic for win or loss):
-    #def results_by_team(team_id, win_loss)
-    result_by_team = @game_teams.values_at(:game_id, :team_id, :result).find_all do |game|
-      game[1] == team_id && game[2] == "WIN"
-    end
-  end
+  #   # Option with full data set, but does not work with current dummy data
+  #   # (this could be made dynamic for win or loss):
+  #   #def results_by_team(team_id, win_loss)
+  #   result_by_team = @game_teams.values_at(:game_id, :team_id, :result).find_all do |game|
+  #     game[1] == team_id && game[2] == "WIN"
+  #   end
+  # end
 
-  def games_by_team(team_id) # List of every game a team played - helper method for issue #18
-    # [[2013020252, 16], [2013020987, 16], [2014020903, 16], [2012020574, 16], [2014030161, 16],
-    games = []
-    @game_teams.each do |row|
-      games <<[row[:game_id], team_id] if (row[:team_id] == team_id)
-    end
-    games
+  # def games_by_team(team_id) # List of every game a team played - helper method for issue #18
+  #   # [[2013020252, 16], [2013020987, 16], [2014020903, 16], [2012020574, 16], [2014030161, 16],
+  #   games = []
+  #   @game_teams.each do |row|
+  #     games <<[row[:game_id], team_id] if (row[:team_id] == team_id)
+  #   end
+  #   games
 
-    # Option with full data set, but does not work with current dummy data:
-    # games_by_team = @game_teams.values_at(:game_id, :team_id, :result).find_all do |game|
-    #   game[1] == team_id
-    # end
-  end
+  #   # Option with full data set, but does not work with current dummy data:
+  #   # games_by_team = @game_teams.values_at(:game_id, :team_id, :result).find_all do |game|
+  #   #   game[1] == team_id
+  #   # end
+  # end
 
-  def number_team_games_per_season(team_id) # Count of number of games a team played each season - helper method for issue #18
-    # {20122013=>1, 20132014=>2, 20142015=>7, 20162017=>4}
-    team_games_by_season = Hash.new{0}
-    games_by_season.each do |season, games|
-      games_by_team(team_id).each do |result_data|
-        if games.include?(result_data[0])
-          team_games_by_season[season] += 1.0
-        elsif team_games_by_season[season] == 0
-          team_games_by_season[season] = 0.0
-        end
-      end
-    end
-    team_games_by_season
-  end
+  # def number_team_games_per_season(team_id) # Count of number of games a team played each season - helper method for issue #18
+  #   # {20122013=>1, 20132014=>2, 20142015=>7, 20162017=>4}
+  #   team_games_by_season = Hash.new{0}
+  #   games_by_season.each do |season, games|
+  #     games_by_team(team_id).each do |result_data|
+  #       if games.include?(result_data[0])
+  #         team_games_by_season[season] += 1.0
+  #       elsif team_games_by_season[season] == 0
+  #         team_games_by_season[season] = 0.0
+  #       end
+  #     end
+  #   end
+  #   team_games_by_season
+  # end
 
-  def number_team_wins_per_season(team_id) # Count of number of games a team won each season -helper method for issue #18
-    # {20132014=>1, 20142015=>1, 20162017=>2}
-    wins_by_season = Hash.new{0}
-    games_by_season.each do |season, games|
-      wins_by_team(team_id).each do |result_data|
-        if games.include?(result_data[0])
-          wins_by_season[season] += 1.0
-        end
-      end
-      if wins_by_season[season] == 0
-        wins_by_season[season] = 0.0
-      end
-    end
-    wins_by_season
-  end
+  # def number_team_wins_per_season(team_id) # Count of number of games a team won each season -helper method for issue #18
+  #   # {20132014=>1, 20142015=>1, 20162017=>2}
+  #   wins_by_season = Hash.new{0}
+  #   games_by_season.each do |season, games|
+  #     wins_by_team(team_id).each do |result_data|
+  #       if games.include?(result_data[0])
+  #         wins_by_season[season] += 1.0
+  #       end
+  #     end
+  #     if wins_by_season[season] == 0
+  #       wins_by_season[season] = 0.0
+  #     end
+  #   end
+  #   wins_by_season
+  # end
 
-  def season_win_percentage(team_id) # Percentage of won games per season by team - helper method for issue #18
-    # {20132014=>50.0, 20142015=>14.3, 20162017=>50.0}
-    win_percentage = Hash.new
-    number_team_games_per_season(team_id).each do |game_season, game_count|
-      number_team_wins_per_season(team_id).each do |wins_season, win_count|
-        if game_season == wins_season
-          if win_count == 0
-            percentage = 0.0
-          else
-            percentage = ((win_count/game_count.to_f) * 100).round(1)
-          end
-          win_percentage[game_season] = percentage
-        end
-      end
-    end
-    win_percentage
-  end
+  # def season_win_percentage(team_id) # Percentage of won games per season by team - helper method for issue #18
+  #   # {20132014=>50.0, 20142015=>14.3, 20162017=>50.0}
+  #   win_percentage = Hash.new
+  #   number_team_games_per_season(team_id).each do |game_season, game_count|
+  #     number_team_wins_per_season(team_id).each do |wins_season, win_count|
+  #       if game_season == wins_season
+  #         if win_count == 0
+  #           percentage = 0.0
+  #         else
+  #           percentage = ((win_count/game_count.to_f) * 100).round(1)
+  #         end
+  #         win_percentage[game_season] = percentage
+  #       end
+  #     end
+  #   end
+  #   win_percentage
+  # end
 
   def best_season (team_id) #issue # 18 - PASS
     season_win_percentage(team_id.to_i).key(season_win_percentage(team_id.to_i).values.max).to_s
